Refactor bot into multiple files
