+++ bot/__init__.py
@@ -1,0 +1,1 @@
+# Bot package
