+++ bot/config.py
@@ -1,0 +1,1 @@
+BOT_TOKEN = "7225094617:AAHY3aLYmIiwLZKcGXx9em3ULK7eviko1NU"
